requirements.txt
mcp[cli]
python-dotenv
uvicorn
pydantic
fastapi
